+++ Weapons/Sharpness.cs
@@ -2,7 +2,7 @@
 
 namespace MHW_Editor.Weapons {
     public partial class Sharpness : MhwItem {
-        public Sharpness(byte[] bytes, int offset) : base(bytes, offset) {
+        public Sharpness(byte[] bytes, ulong offset) : base(bytes, offset) {
         }
 
         public override string Name => "None";
