+++ Weapons/ShellTable.cs
@@ -2,11 +2,11 @@
 
 namespace MHW_Editor.Weapons {
     public partial class ShellTable : MhwItem {
-        public ShellTable(byte[] bytes, int offset) : base(bytes, offset) {
+        public ShellTable(byte[] bytes, ulong offset) : base(bytes, offset) {
         }
 
         public override string Name => "None";
 
-        public int Index => (Offset - InitialOffset) / StructSize;
+        public ulong Index => (Offset - InitialOffset) / StructSize;
     }
 }
